Move bokeh layout setup into class
Moving the layout setup code into a class to eliminate use of global
variables.  When using the "standalone embed" approach (as opposed to
"bokeh serve"), it is necessary to eliminate global bokeh "models",
otherwise clicking refresh in the browser will cause a crash with
"models must be owned by only a single document".  See also
https://github.com/bokeh/bokeh/issues/8265.