Fix application entry point in setup.py
With this fix, it is now possible to install and then launch the app
from any directory.  From the source directory, run:
python setup.py develop --user
The "develop" command installs a link to the source code instead of
actually copying the files as would be done in a true install.  After
this, the command "para_atm" can be run from any directory and should
launch the app.

At this point, the relative paths used by things like readNATS will
still work when launching "para_atm" with a "development install",
since the source files are still in the same place relative to NATS
and the other directories.  Path treatment will need to be reworked
for handling real installations.