Fixing a whole host of errors related to reading from db
Various bugs that were preventing the app from actually reading data
from the database when it was available.  Instead, selecting an
existing table from the menu ended up re-reading from the NATS file.
This happened because overly broad exception handling was masking a
TypeError that was arising due to a bug in how getNATSdata was being
called.

The fix uses more specific exception handling (although longer
term this approach for identifying the type of data in the database
should be completely reworked).  However, that identified a second
problem in LaunchApp: the NATS-specific handling of results['time']
was not behaving correctly when reading from the database.  As a
temporary fix, I moved that check outside of the "except" block so
that gets done regardless of whether data are read from the db or NATS
File.  However, this won't be compatible with removing the NATS file
after it has been read.  This whole approach should be reworked.