Add test for table selection change
When the table selection changes, the app automatically tries to read
in the indicated table from the database.  To test this, we first read
the sample NATS file (now possible via the general
read_nats_output_file function) and add it to the database as
'test_table'.  The tables selection is changed to trigger the
callback.  Lastly, the temporary table is removed.