Reduce memory used by read_iff_file
Previously, it crashed my linux VM when trying to read a file with
1,020,410 record type 3 entires.  Appears that a big contributor was
converting the raw lines into a np.array object, which required about
5x more memory than just storing as a list.  Switching to list
necessitated only a small change to how the subset of lines is
extracted.

Also switching to low_memory=True in read_csv.  We can try this (which
is the default) and see whether it creates issues in inferring the
data types.