Refactoring code for ground_ssd calculation
Reorganizing into function ground_ssd_safety_analysis that accepts a
DataFrame and returns a DataFrame.  Added function to infer status for
speed if status is not present in input.

It does run, but have not checked the results, and there are some
numpy array warnings that need to be cleaned up.