Initial implementation of NatsSimulationWrapper class
This class is intended to transparently handle setting up and running
NATS simulations.  It handles all path issues related to (1) Java
classpath, (2) changing current directory so that "./run" will work,
and (3) output files.  It also handles JVM startup and shutdown.  The
user simply derives from the class and implements simulation,
write_output, and cleanup methods.

Also there is a simple method get_nats_constant that can dynamically
access named NATS constants, avoiding the need to explicitly declare
them (as in NATS_Python_Header).

Next steps will be to extend the class so that it behaves like a
function with user-specifeid variables that get passed to the
simulation method.

Adding jpype1==0.6.3 as dependency.