Fixes to support natsStandalone.stop() calls
Adding a missing natsStandalone.stop() call to the cleanup method in
the nats_gate_to_gate test case.  With this change, also discovered
that the command to change back to the starting directory needs to
happen after the call to natsStandalone.stop(), otherwise it leads to
crashes or hangs in Java.