Remove dbSetup.sh script
This was used to install database tools on the system (assuming Ubuntu
Linux) and to import a backup database.  At this time, the backup
database is not currently included in the repository.  The current
vision for PARA_ATM does not include this particular functionality.
It may be revisited in the future if needed.

Also updating README to remove reference to database installation.