Rework read_iff_file to reduce memory usage and add callsigns option
Do not read entire file into memory to determine record types.
Instead, this is done with a list comprehension on the file iterator.

Add callsigns option, which can be used to specify which aircraft
callsigns to extract from the IFF file.  This may be useful for
extracting a portion out of a very large IFF file.

Use chunking in pd.read_csv, so as not to read the entire file into a
data frame.  Combined with the option to specify callsigns, this makes
it possible to read very large files into a data frame with modest
memory usage.

Also add encoding option and make default 'latin-1'.  This was found
to be necessary to read one particular IFF file that apparently
contains an incorrect byte, which raises an exception with the default
encoding.