Fix NATS simulation PAUSE issues in VCAS and re-enable test
See GitHub Issue #5.  Previously, the NATS simulation would sometimes
"miss" the PAUSE's and run to the end, which caused the Python code to
stick in a perpetual for loop, because the simulation status was END
and the code only checks for PAUSE.

The solution was to add sleep statements inside the for loop, as is
done in the examples that are included with NATS.

I have also re-enabled the VCAS unittest.