Simplify read_nats_output_file to call read_gnats_output_file
Logic is the same between the two, but read_nats_output_file had a
leftover filter to only retain a subset of columns, which was later
dropped from the read_gnats_output_file.  For consistency, changed so
that read_nats_output_file just calls read_gnats_output_file.  Update
test results due to the additional columns that are now returned.